Switch to the pdf crate
This crate is faster than the lopdf crate. However, it creates a PDF
file where text cannot be selected, which is bad. This commit is a
checkpoint in case I want to go back to this crate. I am going to
try the mupdf crate next to see if it is better.